// patientorfullstack/patientor-backend/src/types.ts
export enum Gender {
  Male = "male",
  Female = "female",
}

export enum HealthCheckRating {
  "Healthy" = 0,
  "LowRisk" = 1,
  "HighRisk" = 2,
  "CriticalRisk" = 3
}

export interface Diagnosis {
  code: string;
  name: string;
  latin?: string;
}

interface BaseEntry {
  id: string;
  description: string;
  date: string;
  specialist: string;
  diagnosisCodes?: Array<Diagnosis['code']>
}

export interface HealthCheckEntry extends BaseEntry {
  type: "HealthCheck";
  healthCheckRating: HealthCheckRating;
}

export type Discharge = {
  date: string;
  criteria: string;
};

export type SickLeave = {
  startDate: string;
  endDate: string;
};

export interface OccupationalHealthcareEntry extends BaseEntry {
  type: "OccupationalHealthcare";
  employerName: string;
  sickLeave?: SickLeave
}

export interface HospitalEntry extends BaseEntry {
  type: "Hospital",
  discharge: Discharge
}

export interface Patient {
  id: string;
  name: string;
  dateOfBirth: string;
  ssn: string;
  gender: Gender;
  occupation: string;
  entries: Entry[]
} 

export type NonSensitivePatientEntry = Omit<Patient, 'ssn' | 'entries'>;

export type NonSensitiveDiagnoseEntry = Omit<Diagnosis, "latin">;

export type newPatientEntry = Omit<Patient, "id">;

export type HealthCheckEntryWithoutId = Omit<HealthCheckEntry, "id">;

export type HospitalEntryWithoutId = Omit<HospitalEntry, "id">;

export type OccupationalHealthcareEntryWithoutId = Omit<OccupationalHealthcareEntry, "id">;

export type Entry =
  | HospitalEntry
  | OccupationalHealthcareEntry
  | HealthCheckEntry;

type UnionOmit<T, K extends string | number | symbol> = T extends unknown ? Omit<T, K> : never;

export type EntryWithoutId = UnionOmit<Entry, 'id'>;